requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==3.0.2
psutil==6.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
structlog==24.4.0
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.2
zipp==3.20.2
